+++ docs/examples-raw/contrib_tables_add_update.py
@@ -7,3 +7,9 @@ assert list(
     .update(c=c.call_func(abs, c.col("c")))  # UPDATING NEW COLUMN: "c"
     .into_iter_rows(dict)
 ) == [{"a": 1, "b": -2, "c": 1}, {"a": 2, "b": -3, "c": 1}]
+
+assert list(
+    Table.from_rows([("a", "b"), ("xy", "zw")], header=True)
+    .update_all(c.this + c.col("b"))
+    .into_iter_rows(dict)
+) == [{"a": "xyzw", "b": "zwzw"}]